# internal/delivery/docs/swagger.yaml
info:
  contact: {}
paths:
  /chat:
    post:
      consumes:
      - application/json
      parameters:
      - description: Chat name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created chat with id
          schema:
            type: integer
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create chat
      tags:
      - chat
  /chat/by_page:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully returned chats
          schema:
            type: integer
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get chats by page
      tags:
      - chat
  /chat/messages:
    get:
      consumes:
      - application/json
      parameters:
      - description: Chat id
        in: query
        name: chat_id
        required: true
        type: string
      - description: Page
        in: query
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully returned messages
          schema:
            type: integer
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get chat messages
      tags:
      - chat
swagger: "2.0"